+++ src/pages/AllProducts.tsx
@@ -1,13 +1,13 @@
-import { FC, useEffect, useRef, useState } from "react";
+import { FC, useEffect, useState } from "react";
 import { useAppSelector, useAppDispatch } from "../redux/hooks";
 import { addProducts } from "../redux/features/productSlice";
 import ProductCard from "../components/ProductCard";
 import { Product } from "../models/Product";
 import { updateLoading } from "../redux/features/homeSlice";
+import SortProducts from "../components/SortProducts"
 
 const AllProducts: FC = () => {
   const dispatch = useAppDispatch();
-  const sortRef = useRef<HTMLSelectElement>(null);
   const [currentProducts, setCurrentProducts] = useState<Product[]>([]);
   const allProducts = useAppSelector(
     (state) => state.productReducer.allProducts
@@ -32,49 +32,15 @@ const AllProducts: FC = () => {
     setCurrentProducts(allProducts);
   }, [allProducts]);
 
-  const sortProducts = (sortValue: string) => {
-    if (sortValue === "asc") {
-      setCurrentProducts(
-        [...currentProducts].sort((a, b) => {
-          const aPrice =
-            a.price - (a.price * (a.discountPercentage ?? 0)) / 100;
-          const bPrice =
-            b.price - (b.price * (b.discountPercentage ?? 0)) / 100;
-          return aPrice - bPrice;
-        })
-      );
-    } else if (sortValue === "desc") {
-      setCurrentProducts(
-        [...currentProducts].sort((a, b) => {
-          const aPrice =
-            a.price - (a.price * (a.discountPercentage ?? 0)) / 100;
-          const bPrice =
-            b.price - (b.price * (b.discountPercentage ?? 0)) / 100;
-          return bPrice - aPrice;
-        })
-      );
-    } else {
-      setCurrentProducts([...currentProducts].sort((a, b) => a.id - b.id));
-    }
-  };
-
   return (
     <div className="container mx-auto min-h-[83vh] p-4 font-karla">
       <div className="grid grid-cols-4 gap-1">
         <div className="col-span-4 space-y-4">
           <div className="flex items-center justify-between">
             <span className="text-lg dark:text-white">Products</span>
-            <select
-              ref={sortRef}
-              className="border border-black dark:border-white rounded p-1 dark:text-white dark:bg-slate-600"
-              onChange={(e) => sortProducts(e.target.value)}
-            >
-              <option value="default">Default</option>
-              <option value="asc">Price (low to high)</option>
-              <option value="desc">Price (high to low)</option>
-            </select>
+            <SortProducts products={currentProducts} onChange={setCurrentProducts} />
           </div>
-          
+
           {isLoading ? (
             <div className="flex items-center justify-center">
               <div className="animate-spin mt-32 rounded-full h-12 w-12 border-t-2 border-b-2 border-gray-900 dark:border-white"></div>
@@ -85,8 +51,8 @@ const AllProducts: FC = () => {
                 <ProductCard key={product.id} {...product} />
               ))}
             </div>
-          )} 
-          
+          )}
+
         </div>
       </div>
     </div>
